Nft view
Hi, this way the display of nfts has changed, each row shows 2 nfts fully. The goal would be to do it like on the web version (i.e. the nft name is written in a white box). however the display of nfts should be definitely better.
Have a nice day and hope I have helped you